+++ limbus/app/templates/sample/index.html
@@ -6,7 +6,6 @@
 
     <div class="container">
 
-
         <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pb-2 mb-3 border-bottom">
             <h2><i class="fas fa-vial"></i> Sample Acquisition Portal</h2>
         </div>
