+++ limbus/app/templates/sample/sample/view.html
@@ -5,13 +5,37 @@
 {% block body %}
     <div class="container">
 
-    <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pb-2 mb-3 border-bottom">
-            <h2><i class="fas fa-vial"></i> LIMBSMP-{{ sample.id }}</h2>
-
+        <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pb-2 mb-3 border-bottom">
+            <h2><i class="fas fa-vial"></i> LIMBSMP-{{ sample["id"] }}</h2>
         </div>
 
+        <div class="btn-toolbar form-group" role="toolbar" aria-label="Toolbar with button groups">
+          <div class="btn-group mr-2" role="group" aria-label="First group">
+            <button type="button" class="btn btn-outline-dark"><i class="fa fa-edit"></i> Edit</button>
+          </div>
+          <div class="btn-group mr-2" role="group" aria-label="Second group">
+            <button type="button" class="btn btn-outline-dark"><i class="fa fa-code-branch"></i> Derive</button>
+          </div>
+          <div class="btn-group mr-2" role="group" aria-label="Third group">
+                <button type="button" class="btn btn-outline-dark"><i class="fa fa-code-branch"></i> Aliquot</button>
+          </div>
+            <div class="btn-group" role="group" aria-label="Third group">
+            <button type="button" class="btn btn-outline-dark"><i class="fa fa-print"></i> Print</button>
+          </div>
+        </div>
 
-
+        <div class="row">
+            <div class="col-6">
+                {{ show_value(sample["biobank_barcode"], "Sample Barcode", "text")  }}
+                {{ show_value(sample["sample_status"], "Sample Status", "text")  }}
+                {{ show_value(sample["collection_date"], "Collection Date", "datetime") }}
+            </div>
+            <div class="col-6">
+                {{ show_value(sample["sample_type"], "Sample Type", "text") }}
+                {{ show_value(sample["quantity"], "Quantity", "text")  }}
+                {{ show_value(sample["current_quantity"], "Current Quantity", "text") }}
+            </div>
+        </div>
 
     </div>
 {% endblock %}
